developer/internal_app/requirements.txt
Flask==3.0.0
psutil==5.9.6
Pillow
pytesseract
